refactor: remame every mention to useRef by useReference in global-hooks package
