+++ packages/@louffee-global-hooks/src/index.ts
@@ -19,8 +19,8 @@ export { default as UseStateTuple } from './useState/UseStateTuple'
 
 export { default as useReducer } from './useReducer/useReducer'
 
-export { default as useRef } from './useRef/useRef'
-export { default as MutableRefObject } from './useRef/MutableRefObject'
+export { default as useRef } from './useReference/useReference'
+export { default as MutableRefObject } from './useReference/MutableReferenceObject'
 
 export { default as useFirstMountState } from './useFirstMountState/useFirstMountState'
