--- packages/@louffee-global-hooks/src/useRef/MutableRefObject.ts
@@ -1,5 +1,0 @@
-interface MutableRefObject<TReferenceValue> {
-  current: TReferenceValue
-}
-
-export default MutableRefObject
