+++ packages/@louffee-global-hooks/src/useReference/MutableReferenceObject.ts
@@ -1,0 +1,5 @@
+interface MutableReferenceObject<TReferenceValue> {
+  current: TReferenceValue
+}
+
+export default MutableReferenceObject
